third_party/blink/web_tests/platform/generic/editing/selection/type-checking-expected.txt

PASS selection.collapse(document.body, 0) did not throw exception.
PASS selection.collapse(null, 0) did not throw exception.
PASS selection.collapse(undefined, 0) did not throw exception.
PASS selection.collapse(0, 0) threw exception TypeError: Failed to execute 'collapse' on 'Selection': parameter 1 is not of type 'Node'..
PASS selection.collapse(window, 0) threw exception TypeError: Failed to execute 'collapse' on 'Selection': parameter 1 is not of type 'Node'..
PASS selection.setPosition(document.body, 0) did not throw exception.
PASS selection.setPosition(null, 0) did not throw exception.
PASS selection.setPosition(undefined, 0) did not throw exception.
PASS selection.setPosition(0, 0) threw exception TypeError: Failed to execute 'setPosition' on 'Selection': parameter 1 is not of type 'Node'..
PASS selection.setPosition(window, 0) threw exception TypeError: Failed to execute 'setPosition' on 'Selection': parameter 1 is not of type 'Node'..
PASS selection.setBaseAndExtent(document.body, 0, document.body, 0) did not throw exception.
PASS selection.setBaseAndExtent(null, 0, null, 0) did not throw exception.
PASS selection.setBaseAndExtent(undefined, 0, undefined, 0) did not throw exception.
PASS selection.setBaseAndExtent(0, 0, null, 0) threw exception TypeError: Failed to execute 'setBaseAndExtent' on 'Selection': parameter 1 is not of type 'Node'..
PASS selection.setBaseAndExtent(null, 0, 0, 0) threw exception TypeError: Failed to execute 'setBaseAndExtent' on 'Selection': parameter 3 is not of type 'Node'..
PASS selection.setBaseAndExtent(window, 0, null, 0) threw exception TypeError: Failed to execute 'setBaseAndExtent' on 'Selection': parameter 1 is not of type 'Node'..
PASS selection.setBaseAndExtent(null, 0, window, 0) threw exception TypeError: Failed to execute 'setBaseAndExtent' on 'Selection': parameter 3 is not of type 'Node'..
PASS successfullyParsed is true

TEST COMPLETE
